Leverage docker-systemctl-replacement
The test suite dockerfiles included a task to download the 
docker-systemctl-replacement application, but it was only actually 
functional on centos-7. Placing it in the proper path on Ubuntu, and 
adding the required Python3 interpreter on Ubuntu and Centos 8 allows us 
to actually use the systemctl replacement, dropping the need for 
privileged containers and extra host volume mappings normally required 
by systemd.